05-pharma-ai/requirements.txt
rdkit-pypi
pandas
numpy<2.0
matplotlib
py3Dmol
biopython
scikit-learn
fastapi
uvicorn
pydantic